fixing python, feature flags
