+++ libafl/src/corpus/mod.rs
@@ -227,7 +227,7 @@ pub mod pybind {
 
     macro_rules! unwrap_me {
         ($wrapper:expr, $name:ident, $body:block) => {
-            crate::unwrap_me_body!(
+            libafl_bolts::unwrap_me_body!(
                 $wrapper,
                 $name,
                 $body,
@@ -244,7 +244,7 @@ pub mod pybind {
 
     macro_rules! unwrap_me_mut {
         ($wrapper:expr, $name:ident, $body:block) => {
-            crate::unwrap_me_mut_body!(
+            libafl_bolts::unwrap_me_mut_body!(
                 $wrapper,
                 $name,
                 $body,
