+++ libafl/src/stages/mod.rs
@@ -380,7 +380,7 @@ pub mod pybind {
 
     macro_rules! unwrap_me_mut {
         ($wrapper:expr, $name:ident, $body:block) => {
-            crate::unwrap_me_mut_body!($wrapper, $name, $body, PythonStageWrapper,
+            libafl_bolts::unwrap_me_mut_body!($wrapper, $name, $body, PythonStageWrapper,
                 { StdMutational },
                 {
                     Python(py_wrapper) => {
