+++ libafl_bolts/Cargo.toml
@@ -13,27 +13,18 @@ categories = ["development-tools::testing", "emulators", "embedded", "os", "no-s
 
 [features]
 default = ["std", "derive", "llmp_compression", "llmp_small_maps", "llmp_broker_timeouts", "rand_trait", "fork", "prelude", "gzip", "regex"]
-std = ["serde_json", "serde_json/std", "hostname", "nix", "serde/std", "bincode", "wait-timeout", "byteorder", "once_cell", "uuid", "tui_monitor", "ctor", "backtrace", "uds", "serial_test"] # print, env, launcher ... support
+std = ["serde_json", "serde_json/std", "hostname", "nix", "serde/std", "bincode", "wait-timeout", "byteorder", "once_cell", "uuid", "ctor", "backtrace", "uds", "serial_test"] # print, env, launcher ... support
 derive = ["libafl_derive"] # provide derive(SerdeAny) macro.
 fork = [] # uses the fork() syscall to spawn children, instead of launching a new command, if supported by the OS (has no effect on Windows, no_std).
 rand_trait = ["rand_core"] # If set, libafl's rand implementations will implement `rand::Rng`
-introspection = [] # Include performance statistics of the fuzzing pipeline
-concolic_mutation = ["z3"] # include a simple concolic mutator based on z3
 python = ["pyo3", "concat-idents"]
 prelude = [] # Expose libafl::prelude for access without additional using directives
-tui_monitor = ["tui", "crossterm"] # enable TuiMonitor with crossterm
-prometheus_monitor = ["std", "async-std", "prometheus-client", "tide", "futures"]
 cli = ["clap"]  # expose bolts::cli for easy commandline parsing
 qemu_cli = ["cli"] # Commandline flags for qemu-based fuzzers
 frida_cli = ["cli"] # Commandline flags for frida-based fuzzers
-afl_exec_sec = [] # calculate exec/sec like AFL
 errors_backtrace = ["backtrace"]
-cmin = ["z3"] # corpus minimisation
-corpus_btreemap = [] # Switches from HashMap to BTreeMap for CorpusId
 gzip = ["miniz_oxide"] # Enables gzip compression in certain parts of the lib
 regex = ["std", "dep:regex"] # enables the NaiveTokenizer and StacktraceObserver
-casr = ["libcasr", "std", "regex"] # enables deduplication based on libcasr for StacktraceObserver
-tcp_manager = ["tokio", "std"] # A simple EventManager proxying everything via TCP
 
 # features hiding dependencies licensed under GPL
 gpl = []
@@ -98,8 +89,6 @@ log = "0.4.18"
 tokio = { version = "1.28.1", optional = true, features = ["sync", "net", "rt", "io-util", "macros"] } # only used for TCP Event Manager right now
 
 wait-timeout = { version = "0.2", optional = true } # used by CommandExecutor to wait for child process
-
-z3 = { version = "0.11", features = ["static-link-z3"], optional = true } # for concolic mutation
 
 pyo3 = { version = "0.18.3", optional = true, features = ["serde", "macros"] }
 concat-idents = { version = "1.1.3", optional = true }
